Move bottom nav to a seperate container. Move Join family as a page
